Finished 0-2-5 Yandex Practicum Rust Course
We have covered patter-mathcing with `match` and such interesting
constructions as: `if let`, `while let`, `let else`. I am debating
whether to start writing a commit message for every lesson completed, to
boost the activity overview graph on Github or not. I am also practicing
using LazyGit with nvim - and so I want to write as many messages as
possible.

I really find the experience with nvim cool. It feels ergonomic. I also
like the idea, that I will feel OK, even if I have to work on a remote
computer, and I can only ssh into it.