Finished 0-2-8
In this section and a couple of previous ones we have explored traits,
generics, ownership, references. The theory provided in the course is
good, like there is a lot of it, but the project which we should develop
on the side is not covering the whole spectrum of the covered material.
Like I do not understand what is the meaning behind the `type Status =
char` thingy, as well as I do not understand what is the meaning behind
the `const ENTITY_TYPE: &'static str = "SmartDevice"` thingy. I should
read the documentation I think.