Finished 0-2-9
In this module we have considered traits Display and Debug, and also
`dbg!` macro was mentioned.

I liked this, it was instructive, but I do not understand what is the
meaning of <'_> inside the type of the parameter &mut fmt::Formatter<'_>
-- does it have something to do with the generics?

On the side I really want to start using the gbd inside the nvim. I feel
so much better in this IDE - it feels cozy, I also do not have copilot
or anything of that sort and I think myself. But whenever I am calling a
function, like when I am writing the call itself - nvim autocompletes it
with the right number of argument, if I type just one character the
placeholder with the name of the parameter disapears, and I can enter
it. But I do not know how to get to the next one. I have to go to the
normal mode, remove the placeholder and then enter the insert mode
again. There has to be a workaraound for that, but I do not know what it
is.