Finished 0-3-2
In this chapter we have covered the `Option<T>` and `Result<T, E` and
the syntactic sugar operator `?`. I understand the concept of Rust - we
want to `panic!` if something failed (actually is not reasonable
anymore), but are required to deal with all the other cases.

I still do not understand the concept of the trait (although it has
something to do with the shared behaviour). Like it was mentioned in the
reading, that `std::error:Error` is a trait, that guarantees the shared
behaviour of the errors from different rust projects. Like if we define
our custom error as:

```
#[derive(Debug, PartialEq)]
enum MathError
```

why does this trait matter, or osmething is happening behind the scenes?