Started 0-3-3
Here we focused on the functional style for dealing with Option and
Result, rather than the imperative one with a bunch of matches as in the
previous lesson. I learned that I can use `.ok()` to convert `Result<T,
E>` into Option<T> and I get Some if the result was Ok, and None if the
Result was an error.

We have also covered the methods: `map`, `map_err` and many others. I am
honestly confused by the number, and I am sure, that I would not be able
to choose the correct one.

We have also talked about implementing the trait Error for the Custom
Error num. And I am a bit confused by the purpuse. It was shown as if
our custom Error has an an option to contain some other error inside.
Why?

I also noticed, that I can use nvim to remind me of the difficult
function signatures for the impl SomeTrait for MyCustomEnum. I just need
to remember, that for Display I need fmt, and for the Error I need the
source.

The learning continues...