Finished 0-3-3
In this lesson we have learned, that `Box<dyn Err>` can be used to
return multiple types of errors, without the need to create a custom
error type, bascially the only thing that is required is the
implementation of the trait Error. We have also learned about the `From`
and `Into` traits that basically help with converting one type of Error
into the other. But there wasn't any practice for them in the lesson. I
am still struggling with a plephora of different Option methods, there
are som many of them, that I am lost. Myabe with practice I would have
more expertise. But I feel stupid. Another thing is that I still cannot
distinguish between the `&str` and String, but I have at least learned,
that I need to use: `.as_str()` to convert from String to &str, and
`.to_string()` in the other direction.

Nvim and clippy were very helpfull, and I start to develop a habbit -
made the code work correctly, and then rely on clippy to make sure, that
my code doesn't look like a toddler wrote it.