Finished 0-3-4
In this lesson we have covered some pointers, which I can use in Rust.
But I still do not understand the motivation why. We have

1. `Box<T>` to store stuff on the heap, and we use it work with data,
which we cannot placeo n the stack, because we do not know about the
size.

2. `Rc<T>` to enable multiple owners.

Then we started to talk about the internal mutability. I remember how we
covered & and &mut, but ` RefCell` are enforced by the runtime and not
the compile time. So there can be panics at runtime.

Then we also talked about the `Rc<RefCell<T>>` for the data that has
many owners, with internal mutability and run-time checks.

Again, I am confused.